requirements.txt
#this file will be filled with all requirements for our project
Flask==3.0.0
mysql-connector-python
